Added check that only one instance of prg is concurrently running

+++ code/Audio_Recorder_01.py
@@ -1,50 +1,65 @@
 import pyaudio
 import wave
 import sys
+import os
 from datetime import datetime
+recordingStatusIndicator = "already_rec.txt"
 
 def main():
     if len(sys.argv) > 1:
-        commandLineParameter = sys.argv[1]
-        form_1 = pyaudio.paInt16 # 16-bit resolution
-        chans = 1 # 1 channel
-        samp_rate = 44100 # 44.1kHz sampling rate
-        chunk = 4096 # 2^12 samples for buffer
-        record_secs = 5 # seconds to record
-       
-        dev_index = int(commandLineParameter)
-        wav_output_filename = get_New_Recording_FileName() # name of .wav file
+        if not os.path.exists(recordingStatusIndicator):
+        # make sure only one instance is recording concurrently    
+            f = open(recordingStatusIndicator, "w")
+            f.close()
 
-        audio = pyaudio.PyAudio() # create pyaudio instantiation
+        # get USB mic interface number
+            commandLineParameter = sys.argv[1]
+
+        # setup pyaudio stream params
+            form_1 = pyaudio.paInt16 # 16-bit resolution
+            chans = 1 # 1 channel
+            samp_rate = 44100 # 44.1kHz sampling rate
+            chunk = 4096 # 2^12 samples for buffer
+            record_secs = 5 # seconds to record
+       
+            dev_index = int(commandLineParameter)
+            wav_output_filename = get_New_Recording_FileName() # name of .wav file
+
+            audio = pyaudio.PyAudio() # create pyaudio instantiation
 
         # create pyaudio stream
-        stream = audio.open(format = form_1,rate = samp_rate,channels = chans, \
+            stream = audio.open(format = form_1,rate = samp_rate,channels = chans, \
                             input_device_index = dev_index,input = True, \
                             frames_per_buffer=chunk)
-        print("Recording into: " + wav_output_filename + " for " + str(record_secs) + " seconds...")
-        frames = []
+            print("Recording into: " + wav_output_filename + " for " + str(record_secs) + " seconds...")
+            frames = []
 
         # loop through stream and append audio chunks to frame array
-        for ii in range(0,int((samp_rate/chunk)*record_secs)):
-            data = stream.read(chunk)
-            frames.append(data)
+            for ii in range(0,int((samp_rate/chunk)*record_secs)):
+                data = stream.read(chunk)
+                frames.append(data)
 
-        print("Finished Recording")
+            print("Finished Recording")
 
         # stop the stream, close it, and terminate the pyaudio instantiation
-        stream.stop_stream()
-        stream.close()
-        audio.terminate()
+            stream.stop_stream()
+            stream.close()
+            audio.terminate()
 
         # save the audio frames as .wav file
-        wavefile = wave.open(wav_output_filename,'wb')
-        wavefile.setnchannels(chans)
-        wavefile.setsampwidth(audio.get_sample_size(form_1))
-        wavefile.setframerate(samp_rate)
-        wavefile.writeframes(b''.join(frames))
-        wavefile.close()
+            wavefile = wave.open(wav_output_filename,'wb')
+            wavefile.setnchannels(chans)
+            wavefile.setsampwidth(audio.get_sample_size(form_1))
+            wavefile.setframerate(samp_rate)
+            wavefile.writeframes(b''.join(frames))
+            wavefile.close()
+
+        # remove currently recording status indicator
+            os.remove(recordingStatusIndicator)
+        else:
+            print("Please provide the USB Mic Device Interface Number as Command Line Parameter")
     else:
-        print("Please provide the USB Mic Device Interface Number as Command Line Parameter")
+        print("Recording in progress...")
         
 def get_New_Recording_FileName():
     fileExt = '.wav'
